requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
ffmpeg-python==0.2.0
filelock==3.15.4
fsspec==2024.6.1
future==1.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.8
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.23.5
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyJWT==2.8.0
pyparsing==3.1.4
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
rsa==4.9
sqlparse==0.4.4
sympy==1.13.2
tiktoken==0.7.0
torch==1.13.1
tqdm==4.66.5
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
